Make it possible to depend on haskell_libraries from statically linked cc targets
Previously, only cc targets that were linked _without_ the linkstatic setting
could depend on the Haskell libraries with foreign exports. Any attempt to
use them in a statically-linked target resulted in a "missing shared object"
error at runtime. The dep_on_haskell cc_tests are linked in shared mode (blaze
default for cc_test), but changing the targets to cc_binary is sufficient to
reproduce the failure.

The CL enables linking of Haskell libraries into static targets and adds a
respective test. The main trick is to avoid linking against shared objects
of cc deps directly when calling GHC to emit the Haskell .so, but instead
link all of them _statically_ into a shared library (which blaze is explicitly
disallowed from compiling as a static library), and make the Haskell target
depend on that. This solves the problem of blaze thinking that it's ok to
simply link all the cc deps directly into the final executable, because every
Haskell .so is accompanied by a self-contained cc shared object.

This of course can introduce some inefficiency when multiple Haskell libs have
common cc dependencies. In that case, the dependencies will get compiled into
each library's private copy of the cc deps. Worse still, linking multiple copies
of a single library statically into a single process might result in some surprises
(e.g. Google flag libraries expect to only be linked in once). Due to that, this
setting is off by default, and can be explicitly opt into by wrapping a Haskell
library in `haskell_static_foreign_library`. This will create an alternative
name for the original library that can only be referenced from targets that expect
the CcInfo provider which will be similar to the one returned by the original target,
with the exception of all cc deps being compiled into self-contained bundles as
described above.